Switch to VDAF-13.
I currently pin to a specific commit in libprio-rs tracking the VDAF-13
implementation, as the implementation is currently ongoing. I am fairly
confident that the only breaking interface changes have already been
implemented: there is a new "application context string", which is
computed as the concatenation of hte current DAP draft in use & the task
ID.

Also, bump the version tag while I'm at it, since I touched it anyway.