Bump MSRV to 1.80.
This is, at time of writing, the "stable minus two releases" that Janus
supports. Also, clean up a few uses of OnceLock -- mostly to replace
with the simpler LazyLock stabilized in 1.80, in one place replacing it
with an async mutex as neither OnceLock nor LazyLock were appropriate as
the initializing function would be async.