Fixup garbage-collection-check clause of put_report_aggregation. (#3415)
* Fixup garbage-collection-check clause of put_report_aggregation.

...and put_leader_report_aggregation. The condition here was reversed,
only allowing a non-expired report aggregation row to be updated by the
put operations. In practice, this wasn't a problem, as in application
code put_report_aggregation was always paired with put_aggregation_job,
and the condition on put_aggregation_job was correct.

* add tests for upserting various objects

Adds datastore level tests that check that we correctly upsert on
conflict when inserting leader report aggregations (both in the init and
failed cases), helper report aggregations, scrubbed reports and
aggregation jobs.

* add comments explaining ON CONFLICT clauses in queries

---------

Co-authored-by: Tim Geoghegan <timg@divviup.org>