Use task time precisions for aggregation jobs Durations and Intervals.
This change means that Intervals must start on, and must have their duration be,
multiples of the task time precision. This means that intervals in the tests are
often changing from numbers such as `[100, 101)` to `[100, 200)`, due to a 100s
time precision.

This appears significant, but is logically the same, as there's still no overlap
with the next interval.

* Adds validate_precision to Duration and Interval
* Add a new Interval-creation method, from_time_with_precision, and use it for batches.
* Because batches now have wider intervals, the garbage collection test now needs to
  proceed past the wider interval to operate.