Fix a fencepost error in Interval::merged_with
The semantics of Interval::merged_with were such that Intervals made via it always extended out 1
second longer than was implied by its documentation. This change appears mostly to inline the
`self.merge(Interval::from_time(time))`, except instead of using an interval of `(time, time+1)`,
it changes to a zero-length interval of `(time, time)` and does not special case that
`Duration::ZERO`.

This adds tests to both `Interval::merge` and `Interval::merged_with` to ensure the semantics now
match expectations.

Note: a future update will add `time_precision` as a parameter to `merged_with`, but that is a
bigger change, and this one can stand on its own.