Move remainder of project to `AsyncAggregator` (#3845)
Adopt the supertrait in the remaining parts of the crate where it's
appropriate. This one is rather noisier than previous PRs because the
most egregious `where` blocks have been expunged already.

We still have some "bare" `prio::vdaf::Aggregator` in the client
libraries, but I don't think it'd be appropriate to use
`AsyncAggregator` there.