Stream `BatchAggregation` rows to DB in groups
If time precision is small, or the task's time interval is large, then a
large number of `BatchAggregation`s may be needed to satisfy a
collection job, whether that's existing ones that contain some prepared
report shares or empty ones that we write to force serialization
conflicts with other database transactions. Doing `try_join_all` on
hundreds of thousands of futures all at once causes a huge number of
allocations, which can cause significant memory pressure due to (1)
allocating all the objects that the futures need and (2) the fact that
many of these futures will race to populate the prepare SQL statement
cache, allocating rather large strings along the way.

A first step toward alleviating that is to use
`TryStreamExt::try_for_each_concurrent` so that we only handle a limited
number of futures at once. We do this in both of the transactions in
`step_collection_job_generic` which put or update `BatchAggregation`s,
and the equivalent code path on the helper side.

The number of futures we'll run at once is configurable. The default is
set to 10,000, which is not particularly scientific. I'm not seeing a
very clear correlation between performance and that figure. But my
thinking is that if ~300,000 futures at once was too much, and if peak
memory usage is linear in the number of futures we run concurrently,
then 10,000 should cut peak memory usage by an order of magnitude, and
then some. That seems like an OK default, and we can tune this more in
deployments.

Part of #3857