Store ping pong transition when polling agg jobs
`prio::vdaf::topology::ping_pong::Transition` is designed so that users
can serialize it instead of `PingPongState`, leading to substantial
storage savings in some scenarios. This is why we have

```rust
ReportAggregationState::LeaderContinue {
    transition: PingPongTransition,
}
```

When we did the async aggregation job stuff, we introduced
`ReportAggregationState::LeaderPoll { state: PingPongState }`, but this
makes it impossible to take advantage of `PingPongTransition`. This
commit splits `LeaderPoll` into `LeaderPollInit` and
`LeaderPollContinue`, which correspond to `LeaderInit` and
`LeaderContinue`. This allows us to enjoy the same storage wins as the
synchronous case does.

[sizes]: https://github.com/divviup/libprio-rs/pull/683/#issuecomment-1687210371