ci-build.yml: make it clear which tests failed (#3890)
When you see that step "test status" failed, you have to go check which
of "test" or "test-docker" actually failed, which can be annoying
because the output from each of those steps is looooooong.

By adding a check for `failure()` to "Upload container logs", we make it
run even if the prior test steps failed, and then we can omit the "test
status" step altogether ([1]).

[1]: https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#failure-with-conditions