`janus_messages`: introduce `MediaType` trait
The `MediaType` trait identifies messages that have an HTTP media type
associated with them, intended for use in `content-type` HTTP headers.
Various types defined in `janus_messages` already had an associated
`MEDIA_TYPE` constant, which is now part of the trait. This means that
elsewhere, we can use a trait bound like `prio::codec::Decode +
MediaType` to generically check that an incoming message has the desired
content-type and then decode it into the desired structure.