`janus_client`: cache HPKE and OHTTP keys
Adds a `janus_core::http::cached_resource` module which GETs an HTTP
resource and then caches it, honoring the `no-cache` and `max-age`
`cache-control` directives. We then use this in `janus_client` to cache
HPKE configurations and OHTTP keys.

This is a forward-port of #3903, which went into `release/0.7`. The
major differences:

- use the `janus_messages::MediaType` trait to figure out the desired
  media type instead of passing it around as an argument
- cache OHTTP keys as well as HPKE keys
- introduce a `FromBytes` trait and use that in `CachedResource` instead
  of `prio::codec::Decode` to gracefully handle `Vec<ohttp::KeyConfig`>
- avoid cloning values where possible

Part of #3159