remove unneeded `Arc` in collection job tests
While looking over #3928, I was wondering why the `CollectionJobReq` in
the test needs to be in an `Arc` and why the type annotation is needed.
Sure enough, we can elide those in the new test as well as in an
existing one and the compiler doesn't gripe. Probably left over from a
refactor of a more complex test that previously used a database
transaction to double check something.