Per-task counters for helper prep errors
Expands `TaskAggregationCounter` so that the leader can report the
various `PrepareError`s that a helper might indicate during handling of
an aggregation job. Columns are added to the `task_aggregation_counters`
table to persist these values in the Janus datastore. And the new
counters are also exposed via the aggregator API so that `divviup-api`
can query them.

Part of #3798