Per-task counters for helper prep errors (#3958)
Expands `TaskAggregationCounter` so that the leader can report the
various `PrepareError`s that a helper might indicate during handling of
an aggregation job. Columns are added to the `task_aggregation_counters`
table to persist these values in the Janus datastore. And the new
counters are also exposed via the aggregator API so that `divviup-api`
can query them.

Adds fields for counters for leader-side prepare errors to the
`TaskAggregationCounter` and the corresponding database table. These
counters aren't wired up yet and thus always report 0, but tihs way we
will only migrate the DB and change the aggregator API once.

Part of #3798