Support aggregate share job deletion (#3974)
Changes:
- Add DELETE request support at aggregate_shares endpoint for cleanup operations,
  to comply with DAP-15 section 4.7.4.
- Change the collection job driver to send DELETE to the helper on the next step
  after the collection job state transitions to Deleted.
- Tolerate failures of the Helper to support DELETE on Aggregate Shares

This shortcuts out of the job stepping if the current job state is Deleted, which
could result in a 5xx, an empty body, who knows! We need to treat it specially. In
the event that we transition to Deleted while stepping, the old path will still
work correctly.

Fixes #3962