Define TaskDuration for time_precision
This change prepares for #4019 by separating the timelike types from TaskProv that remain
defined as seconds -- particularly Time Precision -- as their own type, TaskDuration.

This is a bit of a wonky change, in that it then adds some more add_task_duration/
sub_task_duration methods that get used... it's unclear to me whether we'll keep them
by the end of #4019. We can take a TODO to reassses them, because Duration (and Time,
and by extension Interval) get substantial rewrites coming up.