Define TimePrecision for time_precision (#4157)
This change prepares for #4019 by separating the timelike types from TaskProv that remain
defined as seconds -- particularly Time Precision -- as their own type, TimePrecision.

This is a bit of a wonky change, in that it then adds some more add_time_precision/
sub_time_precision methods that get used... it's unclear to me whether we'll keep them
by the end of #4019. We can take a TODO to reassess them, because Duration (and Time,
and by extension Interval) get substantial rewrites coming up.

Co-authored-by: David Cook <dcook@divviup.org>