Fix Interval::EMPTY round-trip asymmetry
FromSql: PostgreSQL empty range is decoded as Interval::EMPTY, which is good. However,
ToSql: Interval::EMPTY encodes to [1970-01-01 00:00:00, 1970-01-01 00:00:00), which
bit me while I was fixing the rest of this.

We shouldn't ever hit this! But we might as well fix it while we're here.

Also, these timestamps in PostgreSQL are microseconds, not millis.